Initial implementation of single touch support.
Currently, touch should work on Android and iOS devices (tested on both). It also should work with any implementations of the W3C Draft Touch Specification:

https://dvcs.w3.org/hg/webevents/raw-file/tip/touchevents.html

The API has not been tested and probably does not work on Firefox on touch devices.

You can test the touch support by running the drag and drop example on a touch device.

Currently, the API automatically detects touch support by checking for touch APIs. If those APIs are available, then the library will assume that touch is supported and will use touch events. If a browser incorrectly reports that it supports touch events (such as some older versions of Google Chrome), then the implementation will fail (as it will use the touch events).

In addition, the code is currently calling preventDefault in the touchstart event listener. This is to prevent the page from scrolling when the user drags their finder on the screen.

My thinking is that I will change the API so the developer specifies whether or not to use the touch support, something like:

var s = new Stage(canvas, useTouchSupport);

This give the developer more controll and allows them to use more robust detection mechanisms, such as Modernizr in order to determine when to use touch support. i.e.:

var s = new Stage(canvas, Modernizr.touch);

Note that this work maps single touch events to mouse events. However, when accessing the MouseEvent.nativeEvent property when using touch, a Touch instance will be returned, and not a MouseEvent instance (although they share a lot of the same properties).

When running on Android 2.2 devices, adding a second touch point may break touch events. This is an Android issue.