Resolved cyclical dependencies between Stage, Container, and DisplayObject
Had to break the initial goal of leaving the internal code untouched.  The dependencies
between Stage, Container, and DisplayObject cause problems.  The internal changes are
minimal: instead of requiring the dependencies at the top with the rest of the node.js
code, it requires them only when calling for that specific Object.