Resolves Cross Dependencies
Had to break the initial goal of this change for node.js.  The cross-
dependency between the Container, Stage, and DisplayObject required that
the Stage object be included within specific places inside the code.

The changes are minimal and should only rear their heads when utilizing
this library for node.js.

A minor addition of setting the Canvas library to 0.7.2 or Greater, since
newer versions should work just fine with this library.