Fixed a bug related to stage transforms and hit testing. See extended description.
Consider this scenario. You have stage 1 and stage 2 side by side. Both
have DisplayObjects which register for DblClick. In stage 2, you've
transformed the stage so that it's coordinate space is shifted on top
of stage 1's physical space. When you click on Stage 1, stage 2
receives a double click event. When it's testing for a hit, it uses X
and Y, which have had the transform applied. Therefore it thinks the
hit occurred within it's bounds, even though it clearly hasn't. Adding
a simple check for inBounds fixes this.