Fixed mouseover for when top canvas(es) don't enable mouseover, but botttom do.
Go ahead with testMouseOver in more cases than before: instead of just
skipping it (as a "redundant listener") whenever a previous stage was
defined, we now skip it only if at least one of the previous stages has
_mosueOverIntervalID defined.
In the cases where no previous stages have mouseOverIntervalID defined
(found = false), we, in the second lop I added, compute eventTarget by
checking the previous stages. This way a stage enabled with mouseover
which is buried under other stages can find the right canvas to set the
cursor style of (it's the canvas of a previous stage which matches the
event target).