Updated WebGL examples for new Ticker.js location.   Removes some equality coersion from BitmapText.
WebGL examples needed to be updated to point to new Ticker.js location.

BitmapText example was attempting to use an apostrophe in the example text, but the example's bitmap font does not have an apostrophe.

Some instances of == and != were changed to === and !== in BitmapText

TwoStages example was using incorrect name for button image on a case-sensitive file system.