Added Graphics command.bitmap() matrix parameter
Graphics.beginBitmapFill() has a third parameter that is a matrix to adjust the transformation of the fill image.  The bitmap() method of the fill command did not have that parameter.  It seems that it will be picked up by the exec() on the G.Fill() if we just store the matrix on the command object in the bitmap() method which is what we do here.