Single Frame MovieClip adjust
From https://github.com/CreateJS/EaselJS/issues/1048 - "I checked the FPS on LG TV and so it is done from 60 to 20. On a static image. After research, I found that in the advance method in MovieClip class there is a constant check to update the frame. I added a change to check if the MovieClip class total frame is equal to 1 to change it the mode of the MovieClip to a single frame. This increases performance back to 60 FPS." - LiorZar

Has been tested but keep an eye on it - @gskinner might want to double check it.